requirements.txt
requests
google-generativeai
openai